Added Report - financial summary, also changed all numbers to use decimal instead of double

+++ api/BusinessLogic/BudgetModels/BudgetCalculationResult.cs
@@ -5,6 +5,6 @@
         public string Period { get; set; } = string.Empty;
         public int Year { get; set; }
         public int Month { get; set; }
-        public double Spent { get; set; }
+        public decimal Spent { get; set; }
     }
 }
